split wallet stuff from lib.rs
